+++ src/components/Footer/index.ts
@@ -1,0 +1,2 @@
+export { Footer } from './Footer';
+export { FooterProps } from './types';
