--- src/components/PageContainer/Footer/index.ts
@@ -1,2 +1,0 @@
-export { Button } from './Button';
-export { ButtonProps, ButtonOwnProps } from './types';
