+++ src/pages/ProfilePage/ProfilePage.tsx
@@ -4,7 +4,7 @@ import { Button } from '../../components/Button';
 import { InputField } from '../../components/InputField';
 import { Header } from '../../components/Header';
 import { PageContainer } from '../../components/PageContainer';
-import { Footer } from '../../components/PageContainer/Footer/Footer';
+import { Footer } from '../../components/Footer';
 import { Title } from '../../components/Title';
 import { Form } from '../../components/Form';
 import { Avatar } from '../../components/Avatar';
